this is a proper firebase situation
